Changes in Makefile and addition of Doxyfile
Changes in Makefile bring the `make doc` option to buil documentation.
Also, `make clean` will remove generated documentation.

Doxyfile contains the rule to make documentation.
`doxygen` is requiered in order to generate the documentation.
If you are in a Debian based distro, you can run
    sudo apt-get install doxygen

or if you are in an Arch based distro
    sudo pacman -S doxygen

For other distros, go to the respective distro documentation.